# server/swagger.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: "Blog Management System"
  description: |
    A scalable Blog Management System API that enables users to register, verify their email, and
    authenticate using JWT-based authentication (access and refresh tokens). The system supports
    secure password management, including password change and email-based password reset
    functionality.

    Authenticated users can create, update, publish, and delete blog posts, as well as engage with
    content through comments. The platform enforces role-based access control (RBAC), allowing
    administrators to manage system resources while regular users interact with blog content.

    The API provides advanced querying capabilities such as pagination, filtering, sorting, and 
    searching to efficiently retrieve blog data.

    The application follows production-grade backend practices including caching, structured logging,
    automated testing, and comprehensive error handling.

    For scalable and reliable deployment, the system is containerized using Docker and supports load
    balancing, database replication, and CI/CD pipelines
  contact:
    name: Md Ismile Hosen Siam
    email: siam679ihs@gmail.com
servers:
  - description: dev
    url: http://localhost:4000/api/v1
...
paths:
...
  "/articles/{id}":
    get:
      summary: Get article
      description: Retrieve a single article. Only returns published articles and extends it's public comment.
      tags: [Article]
      parameters:
        - $ref: "#/components/parameters/idPathParam"
        - in: query
          name: expand
          schema:
            type: string
            example: author,comments
      responses:
        "200":
          description: Retrieved an article
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                    example: Data retrieved
                  data:
                    allOf:
                      - type: object
                        properties:
                          id:
                            type: string
                            example: 507f1f77bcf86cd799439011
                          title:
                            type: string
                            example: first article
                          body:
                            type: string
                            example: first article body
                          cover:
                            type: string
                            format: uri
                            example: https://s.aws.com/myfirstblog/123.jpg
                        required: [id, title, body]
                      - $ref: "#/components/schemas/Timestamp"
                      - type: object
                        properties:
                          author:
                            $ref: "#/components/schemas/AuthorDetails"
                          comments:
                            type: array
                            items:
                              allOf:
                                - type: object
                                  properties:
                                    id:
                                      type: string
                                      example: 507f1f77bcf86cd799439011
                                    body:
                                      type: string
                                      example: this is a comment
                                    author:
                                      type: string
                                      example: 507f1f77bcf86cd799439011
                                - $ref: "#/components/schemas/Timestamp"
                  links:
                    type: object
                    properties:
                      self:
                        type: string
                        format: uri
                        example: "/articles/507f1f77bcf86cd799439011"
                      author:
                        type: string
                        format: uri
                        example: "/articles/507f1f77bcf86cd799439011/author"
                      comments:
                        type: string
                        format: uri
                        example: "/articles/507f1f77bcf86cd799439011/comments"
    put:
      summary: Create or update article
      description: >
        Create or update an article resource. This operation will create a new article if it does not exist, or update the existing article if it already exists. (only for article owner)
      tags: [Article]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/idPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrPutUpdateArticleRequest"
      responses:
        "200":
          description: Article updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                    example: Successfully updated article
                  data:
                    allOf:
                      - $ref: "#/components/schemas/Article"
                      - type: object
                        properties:
                          author:
                            type: string
                            example: 507f1f77bcf86cd799439011
                  links:
                    type: object
                    properties:
                      self:
                        type: string
                        format: uri
                        example: "/articles/507f1f77bcf86cd799439011"
        "201":
          description: Article created
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    example: 201
                  message:
                    type: string
                    example: Article created successfully
                  data:
                    allOf:
                      - $ref: "#/components/schemas/Article"
                      - type: object
                        properties:
                          author:
                            type: string
                            example: 507f1f77bcf86cd799439011
                  links:
                    type: object
                    properties:
                      self:
                        type: string
                        format: uri
                        example: "/articles/507f1f77bcf86cd799439011"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/ServerError"

    patch:
      summary: Update article
      description: >
        Partially update an article. Only the article owner or an admin should
        be allowed to perform this action.
      tags: [Article]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/idPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchArticleRequest"
      responses:
        "200":
          $ref: "#/components/responses/ArticleUpdated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"

    delete:
      summary: Delete article
      description: >
        Delete an article. Only the article owner or an admin should be allowed
        to perform this action.
      tags: [Article]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/idPathParam"
      responses:
        "204":
          description: Article deleted successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
...
  "/articles/{id}/comments":
    get:
      summary: List article comments
      description: >
        Retrieve public comments for a specific article. Only returns public comments.
      tags: [Article]
      parameters:
        - $ref: "#/components/parameters/idPathParam"
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/limitParam"
      responses:
        "200":
          description: Return a list of comments
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                    example: Data retrieved
                  data:
                    type: array
                    items:
                      allOf:
                        - type: object
                          properties:
                            id:
                              type: string
                              example: 507f1f77bcf86cd799439011
                            body:
                              type: string
                              example: this is a comment
                            author:
                              $ref: "#/components/schemas/AuthorDetails"
                        - $ref: "#/components/schemas/Timestamp"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
                  links:
                    type: object
                    properties:
                      self:
                        type: string
                        format: uri
                        example: "/articles/507f1f77bcf86cd799439011/comments?page=&limit="
                      article:
                        type: string
                        format: uri
                        example: "/articles/507f1f77bcf86cd799439011"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"

    post:
      summary: Create article comment
      description: Create a comment for a specific article. Requires authentication.
      tags: [Article]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/idPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArticleCommentRequest"
      responses:
        "201":
          description: Comment posted
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    example: 201
                  message:
                    type: string
                    example: comment posted
                  data:
                    allOf:
                      - $ref: "#/components/schemas/Comment"
                  links:
                    type: object
                    properties:
                      self:
                        type: string
                        format: uri
                        example: "/articles/507f1f77bcf86cd799439011/comments"
                      article:
                        type: string
                        format: uri
                        example: "/articles/507f1f77bcf86cd799439011"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"

  # ==================================================
  # Comment Routes
  # ==================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

    refreshCookie:
      type: apiKey
      in: cookie
      name: refreshToken
  schemas:
    Article:
      allOf:
        - type: object
          properties:
            id:
              type: string
              example: 507f1f77bcf86cd799439011
            title:
              type: string
              example: first article
            body:
              type: string
              example: first article body
            cover:
              type: string
              format: uri
              example: https://s.aws.com/myfirstblog/123.jpg
            status:
              type: string
              enum: [hidden, published]
              example: published
          required: [id, title, body, status]
        - $ref: "#/components/schemas/Timestamp"
...
    CreateOrPutUpdateArticleRequest:
      type: object
      properties:
        title:
          type: string
          example: first article
        body:
          type: string
          example: first article body
        cover:
          type: string
          format: uri
          example: https://s.aws.com/myfirstblog/123.jpg
      required: [title, body]
    PatchArticleRequest:
      type: object
      properties:
        title:
          type: string
          example: updated title
        body:
          type: string
          example: updated body
        cover:
          type: string
          format: uri
          example: https://s.aws.com/myfirstblog/123.jpg
        status:
          type: string
          enum: [hidden, published]
          example: hidden
    Comment:
      allOf:
        - type: object
          properties:
            id:
              type: string
              example: 507f1f77bcf86cd799439011
            body:
              type: string
              example: first comment
            status:
              type: string
              enum: [hidden, public]
              example: public
            article:
              type: string
              example: 507f1f77bcf86cd799439011
            author:
              type: string
              example: 507f1f77bcf86cd799439011
          required: [id, body, status, article, author]
        - $ref: "#/components/schemas/Timestamp"
    CreateArticleCommentRequest:
      type: object
      properties:
        body:
          type: string
          example: this is a comment
      required: [body]
...
    Timestamp:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [createdAt, updatedAt]
    AuthorDetails:
      type: object
      properties:
        id:
          type: string
          example: 507f1f77bcf86cd799439011
        name:
          type: string
          example: Md Ismile Hosen Siam
      required: [id, name]
    Pagination:
      type: object
      properties:
        page:
          type: integer
          format: int32
          example: 2
        limit:
          type: integer
          format: int32
          example: 10
        next:
          type: integer
          format: int32
          example: 3
        prev:
          type: integer
          format: int32
          example: 1
        totalPage:
          type: integer
          format: int32
          example: 20
        totalItems:
          type: integer
          format: int32
          example: 200
      required: [page, limit, totalPage, totalItems]
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 400
              error:
                type: string
                example: Bad request
              message:
                type: string
                example: Validation failed
              data:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                      example: field name
                    message:
                      type: string
                      example: invalid input
                    in:
                      type: string
                      enum: [path, query, header, cookie, body]
                      example: body
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 500
              error:
                type: string
                example: Internal server error
              message:
                type: string
                example: We are sorry for the inconvenience. Please try again later.
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 401
              error:
                type: string
                example: Unauthorized
              message:
                type: string
                example: Unauthorized.
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 403
              error:
                type: string
                example: Forbidden
              message:
                type: string
                example: You do not have permission.
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 404
              error:
                type: string
                example: Not found
              message:
                type: string
                example: Requested resource not found
...
    ArticleUpdated:
      description: Article updated
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 200
              message:
                type: string
                example: Successfully updated article data
              data:
                allOf:
                  - $ref: "#/components/schemas/Article"
                  - type: object
                    properties:
                      author:
                        type: string
                        example: 507f1f77bcf86cd799439011
              links:
                type: object
                properties:
                  self:
                    type: string
                    format: uri
                    example: "/articles/507f1f77bcf86cd799439011"
  parameters:
    pageParam:
      in: query
      name: page
      description: Current page number
      required: false
      schema:
        type: integer
        format: int32
        default: 1
        example: 2
    limitParam:
      in: query
      name: limit
      description: Number of items to return
      required: false
      schema:
        type: integer
        format: int32
        default: 10
        example: 10
...
    idPathParam:
      in: path
      name: id
      required: true
      schema:
        type: string
        example: 507f1f77bcf86cd799439011